Move transformers test
